+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 =========
 
 
+- Change default behavior of ``SQLClient.transaction()`` to **not** override the current session's ``autoflush`` setting (use ``SQLClient.transaction(autoflush=True)`` instead. (**breaking change**)
+- Add boolean ``autoflush`` option to ``SQLClient.transaction()`` to set session's ``autoflush`` value for the duration of the transaction.
 - Add new ``sqlservice.event`` decorators:
 
   - ``on_init_scalar``